frameworks/langgraph-gas-price-analyst-agent/pyproject.toml
[project]
name = "langgraph-gas-price-analyst-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browserbase>=1.2.0",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.12",
    "langgraph>=0.4.1",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.0",
    "rizaio>=0.11.0",
]
